Allow non-member users to delete only their own messages

+++ controllers/messagesController.js
@@ -47,10 +47,18 @@ async function getAllMessages(req, res) {
 
 const deleteMessage = [
     userAuth.redirectIfUnauthenticated,
-    userAuth.redirectIfNotMember,
     async (req, res) => {
-        await queries.deleteMessage(req.params.messageId);
-        res.redirect("/?messageDeleted=true");
+        const { messageId } = req.params;
+        const message = await queries.getMessageById(messageId);
+        
+        // allow members to delete any message
+        // allow users to delete their own messages even if they are not members
+        if (req.user.isMember || req.user.id === message.user.id) {
+            await queries.deleteMessage(req.params.messageId);
+            res.redirect("/?messageDeleted=true");
+        } else {
+            res.redirect("/");
+        }
     },
 ];
